usuario/registrar/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bcrypt=3.2.0=py39h2bbff1b_1
ca-certificates=2022.12.7=h5b45459_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py39h2bbff1b_3
click=8.1.3=win_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
openssl=1.1.1t=h2bbff1b_0
pip=22.3.1=py39haa95532_0
pycparser=2.21=pyhd3eb1b0_0
python=3.9.16=h6244533_0
python-dotenv=0.21.1=pyhd8ed1ab_0
setuptools=65.6.3=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h2bbff1b_0
tzdata=2022g=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py39haa95532_2
